pyproject.toml
[project]
name = "code-rag-mcp"
version = "0.1.0"
description = "MCP RAG server for codebase knowledge — index any GitHub org"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0",
    "lancedb>=0.6",
    "sentence-transformers>=2.2",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "pre-commit"]

[tool.ruff]
target-version = "py312"
line-length = 120
# Throwaway debug-session artifacts (repro scripts, one-off analysis) are not
# maintained code — lint them out of the hook entirely.
extend-exclude = [".claude/debug"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator (less readable for complex conditions)
    "B008",   # function call in default argument (used in MCP tool signatures)
    "RUF001", # ambiguous unicode in strings — Cyrillic is intentional (uk project)
    "RUF002", # ambiguous unicode in docstrings — Cyrillic is intentional (uk project)
    "RUF003", # ambiguous unicode in comments — Cyrillic is intentional (uk project)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: end-to-end tests against real knowledge.db",
]
